_archive/legacy-streamlit/requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
# python -m streamlit run Gestion-de-projet-App.py